Set HTML Boilerplate, title, header, created link
To start my project i coded the HTML Boilerplate and added a title
to let it show up in my tab and chose a header for this main page, which
will contain an image and links to three other recipes. As well I added
the first link, which leads to the first recipe.